backend/requirements.txt
# FastAPI and web server dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
websockets>=12.0

# LightRAG and AI dependencies
lightrag-hku>=1.3.7
openai>=1.51.0
tiktoken>=0.7.0

# Document processing
python-docx>=1.1.0
PyPDF2>=3.0.1
pdfplumber>=0.10.0

# Database and storage
neo4j>=5.14.0
redis>=5.0.1
chromadb>=0.4.17

# Vector embeddings and ML
numpy>=1.26.4
sentence-transformers>=2.2.2
transformers>=4.35.2

# Data processing
pandas>=2.1.3
networkx>=3.2.1

# Environment and configuration
python-dotenv>=1.0.1
pydantic>=2.5.0
pydantic-settings>=2.0.3

# Logging and monitoring
loguru>=0.7.2

# Development and testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.2

# CORS and security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4 